# public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Drachenboot Manager API",
        "version": "2.5.0",
        "description": "API documentation for the Drachenboot Manager application with multi-team support."
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/stripe/setup-intent": {
            "post": {
                "summary": "Create Setup Intent",
                "description": "Creates a Stripe SetupIntent for setting up a new payment method. Creates a Stripe customer if one doesn't exist for the team.",
                "tags": [
                    "Billing"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "teamId": {
                                        "type": "string",
                                        "description": "Team ID to setup payment for"
                                    }
                                },
                                "required": [
                                    "teamId"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Setup intent created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "clientSecret": {
                                            "type": "string",
                                            "description": "Client secret for Stripe Elements"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Team ID is required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Insufficient permissions - must be captain or billing user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/stripe/invoices": {
            "get": {
                "summary": "Get Team Invoices",
                "description": "Retrieves the list of invoices for a team's Stripe customer. Billing user or captain required.",
                "tags": [
                    "Billing"
                ],
                "parameters": [
                    {
                        "name": "teamId",
                        "in": "query",
                        "required": true,
                        "description": "Team ID to fetch invoices for",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of invoices",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "invoices": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": {
                                                        "type": "string"
                                                    },
                                                    "date": {
                                                        "type": "integer",
                                                        "description": "Unix timestamp"
                                                    },
                                                    "amount": {
                                                        "type": "integer",
                                                        "description": "Amount in cents"
                                                    },
                                                    "currency": {
                                                        "type": "string"
                                                    },
                                                    "status": {
                                                        "type": "string"
                                                    },
                                                    "invoicePdf": {
                                                        "type": "string",
                                                        "nullable": true,
                                                        "description": "URL to download PDF"
                                                    },
                                                    "number": {
                                                        "type": "string",
                                                        "nullable": true
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Team ID is required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Insufficient permissions - must be billing user or captain",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Error": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string",
                        "description": "Error message"
                    }
                },
                "required": [
                    "error"
                ]
            }
        },
        "securitySchemes": {
            "sessionAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "authjs.session-token",
                "description": "Session-based authentication using NextAuth.js. Requires active login session."
            }
        }
    },
    "security": [
        {
            "sessionAuth": []
        }
    ],
...
}